I have performed basic tasks. 
Performance has been increased more then 2.35x

| Name of Test | base | optimized | optimized2 | optimized | optimized2 |
| --- | --- | --- | --- | --- | --- |
|  |  |  |  | (x-factor) | (x-factor) |
| Thinning::Size_Only::640x480 | 506.173 ms | 267.613 ms | 214.958 ms | 1.89 | 2.35 |
| Thinning::Size_Only::1280x720 | 1241.078 ms | 703.643 ms | 536.966 ms | 1.76 | 2.31 |
| Thinning::Size_Only::1920x1080 | 3687.893 ms | 2066.198 ms | 1702.525 ms | 1.78 | 2.17 |